renamed ConnectionPool to ConnectionManager. people seems confused by name too much. pool is just one of thing it does, it owns connection management in general.
